test on macos and windows (#43)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

support windows and macos.

## Solution Description
<!-- Please clearly and concisely describe your solution. -->
On windows and macos, pylibseekdb will not be installed and
EmbeddedClient is disabled.

GitHub Action doesn't support setup docker on MacOS and execute linux
container on Windows, so there's no CI cases for windows and macos.