support higher version numpy (#44)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
Support higher version numpy.

We can't install Numpy version 1.26 on Windows directly and it will be
compiled on windows.